update checkhit.gpp/追加 ３Dの当たり判定制作
